Fixes UI bugs

+++ frontend/src/app/pages/comment/comment.component.ts
@@ -9,6 +9,7 @@ import { ErrorResponse } from 'src/app/shared/models/response/ErrorResponse.mode
 import { GenericResponse } from 'src/app/shared/models/response/GenericResponse.model';
 import { CommentService } from 'src/app/shared/services/comment.service';
 import { TokenStorageService } from 'src/app/shared/services/token-storage.service';
+import Swal from 'sweetalert2';
 import { CommentDialogComponent } from './comment-dialog/comment-dialog.component';
 
 @Component({
@@ -116,50 +117,64 @@ export class CommentComponent implements OnInit {
   }
 
   deleteComment(commentId: Number) {
-    this.isLoading = true;
-    const newComments = this.comments.filter(
-      (comment) => comment.commentId !== commentId
-    );
-    this.comments = newComments;
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+    }).then((result) => {
+      if (result.isConfirmed) {
 
-    this.commentService.deleteComment(commentId).subscribe(
-      (res: GenericResponse) => {
-        this.notify.emit({
-          icon: 'success',
-          iconColor: 'white',
-          text: `${res.message}`,
-          background: '#a5dc86',
-          color: 'white',
-        });
-      },
-      (errorResponse: ErrorResponse) => {
-        this.notify.emit({
-          icon: 'error',
-          iconColor: 'white',
-          text: `${errorResponse.message}`,
-          background: '#f27474',
-          color: 'white',
-        });
-      },
-      () => {
-        this.isLoading = false;
+        this.isLoading = true;
+        const newComments = this.comments.filter(
+          (comment) => comment.commentId !== commentId
+        );
+        this.comments = newComments;
+
+        this.commentService.deleteComment(commentId).subscribe(
+          (res: GenericResponse) => {
+            this.isLoading = false;
+            this.notify.emit({
+              icon: 'success',
+              iconColor: 'white',
+              text: `${res.message}`,
+              background: '#a5dc86',
+              color: 'white',
+            });
+          },
+          (errorResponse: ErrorResponse) => {
+            this.isLoading = false;
+            this.notify.emit({
+              icon: 'error',
+              iconColor: 'white',
+              text: `${errorResponse.message}`,
+              background: '#f27474',
+              color: 'white',
+            });
+          }
+        );
       }
-    );
+    });
+
   }
 
   openEditCommentDialog(commentId: Number) {
     this.isLoading = true;
     this.commentService.getComment(commentId).subscribe(
       (comment: CommentResponse) => {
+        this.isLoading = false;
         const dialogRef = this.commentDialog.open(CommentDialogComponent, {
           data: comment,
         });
         dialogRef.afterClosed().subscribe((dataComment: CommentResponse) => {
-          this.isLoading = true;
           this.updateComment(dataComment);
         });
       },
       (errorResponse: ErrorResponse) => {
+        this.isLoading = false;
         this.notify.emit({
           icon: 'error',
           iconColor: 'white',
@@ -167,15 +182,13 @@ export class CommentComponent implements OnInit {
           background: '#f27474',
           color: 'white',
         });
-      },
-      () => {
-        this.isLoading = false;
       }
     );
   }
 
   updateComment(updatedComment: CommentResponse) {
     if (!updatedComment) return;
+    this.isLoading = true;
     const index = this.comments.findIndex(
       (c) => c.commentId === updatedComment.commentId
     );
